chore: Deprecate BooleanColumn
