+++ packages/tables/docs/03-columns/03-icon.md
@@ -51,3 +51,39 @@ IconColumn::make('is_featured')
         'success' => 'published',
     ])
 ```
+
+## Handling booleans
+
+Icon columns can display a check or cross icon based on the contents of the database column, either true or false, using the `boolean()` method:
+
+```php
+use Filament\Tables\Columns\IconColumn;
+
+IconColumn::make('is_featured')
+```
+
+### Customizing the boolean icons
+
+You may customize the icon representing each state. Icons are the name of a Blade component present. By default, [Heroicons](https://heroicons.com) are installed:
+
+```php
+use Filament\Tables\Columns\IconColumn;
+
+IconColumn::make('is_featured')
+    ->boolean()
+    ->trueIcon('heroicon-o-badge-check')
+    ->falseIcon('heroicon-o-x-circle')
+```
+
+### Customizing the boolean colors
+
+You may customize the icon color representing each state. These may be either `primary`, `secondary`, `success`, `warning` or `danger`:
+
+```php
+use Filament\Tables\Columns\IconColumn;
+
+IconColumn::make('is_featured')
+    ->boolean()
+    ->trueColor('primary')
+    ->falseColor('warning')
+```
